refactor: 엔티티 내 중복되는 createdAt, updatedAt 컬럼을 MappedSuperclass로 분리
